f*#$&(*ing floats, man.
The maximum value of a float such that (int)(((float)x) + 1) == x + 1
for integer x is around a million.

Yeah.

This is causing lots of off-by-one errors in the lua side (remember, all
numbers are floats now in lua), especially in constructing walls, etc.

So standardize on doubles.  It's this or the fixed-point insanity (LOL!)

+++ source/client/renderer/camera.h
@@ -16,7 +16,7 @@ class Camera
       Camera(H3DNode parent, const char* name);
       virtual ~Camera();
 
-      void OrbitPointBy(const csg::Point3f &point, float xRot, float yRot, float maxX, float minX);
+      void OrbitPointBy(const csg::Point3f &point, double xRot, double yRot, double maxX, double minX);
       void LookAt(const csg::Point3f &point);
 
       H3DNode GetNode() const;
