+++ source/csg/lua/lua_csg.cpp
@@ -90,10 +90,10 @@ void csg::RegisterLuaTypes(lua_State* L)
             LuaRay::RegisterLuaTypes(L),
             LuaRandomNumberGenerator::RegisterLuaTypes(L),
             LuaMesh::RegisterLuaTypes(L),
-            def("get_rect_centroid", (csg::Point2f(*)(csg::Rect2f const&)) &csg::GetCentroid<float, 2>),
-            def("get_cube_centroid", (csg::Point3f(*)(csg::Cube3f const&)) &csg::GetCentroid<float, 3>),
-            def("get_region_centroid", (csg::Point2f(*)(csg::Region2f const&)) &csg::GetCentroid<float, 2>),
-            def("get_region_centroid", (csg::Point3f(*)(csg::Region3f const&)) &csg::GetCentroid<float, 3>)
+            def("get_rect_centroid", (csg::Point2f(*)(csg::Rect2f const&)) &csg::GetCentroid<double, 2>),
+            def("get_cube_centroid", (csg::Point3f(*)(csg::Cube3f const&)) &csg::GetCentroid<double, 3>),
+            def("get_region_centroid", (csg::Point2f(*)(csg::Region2f const&)) &csg::GetCentroid<double, 2>),
+            def("get_region_centroid", (csg::Point3f(*)(csg::Region3f const&)) &csg::GetCentroid<double, 3>)
          ]
       ]
    ];
