+++ source/csg/point.cpp
@@ -32,53 +32,53 @@ const Point1f Point1f::one(1);
 
 
 template <int C>
-Point<float, C> ToFloatFn(Point<int, C> const& pt);
+Point<double, C> ToFloatFn(Point<int, C> const& pt);
 
 template <>
-Point<float, 1> ToFloatFn(Point<int, 1> const& pt) {
-   return Point<float, 1>(static_cast<float>(pt.x));
+Point<double, 1> ToFloatFn(Point<int, 1> const& pt) {
+   return Point<double, 1>(static_cast<double>(pt.x));
 }
 
 template <>
-Point<float, 2> ToFloatFn(Point<int, 2> const& pt) {
-   return Point<float, 2>(static_cast<float>(pt.x),
-                          static_cast<float>(pt.y));
+Point<double, 2> ToFloatFn(Point<int, 2> const& pt) {
+   return Point<double, 2>(static_cast<double>(pt.x),
+                          static_cast<double>(pt.y));
 }
 
 template <>
-Point<float, 3> ToFloatFn(Point<int, 3> const& pt) {
-   return Point<float, 3>(static_cast<float>(pt.x),
-                          static_cast<float>(pt.y),
-                          static_cast<float>(pt.z));
+Point<double, 3> ToFloatFn(Point<int, 3> const& pt) {
+   return Point<double, 3>(static_cast<double>(pt.x),
+                          static_cast<double>(pt.y),
+                          static_cast<double>(pt.z));
 }
 
 template <int C>
-Point<float, C> csg::ToFloat(Point<int, C> const& pt) {
+Point<double, C> csg::ToFloat(Point<int, C> const& pt) {
    return ToFloatFn(pt);
 }
 
 template <int C>
-Point<float, C> const& csg::ToFloat(Point<float, C> const& pt) {
+Point<double, C> const& csg::ToFloat(Point<double, C> const& pt) {
    return pt;
 }
 
 template <int C>
-Point<int, C> ToIntFn(Point<float, C> const& pt);
+Point<int, C> ToIntFn(Point<double, C> const& pt);
 
 template <>
-Point<int, 1> ToIntFn(Point<float, 1> const& pt)
+Point<int, 1> ToIntFn(Point<double, 1> const& pt)
 {
    return Point<int, 1>(ToInt(pt.x));
 }
 
 template <>
-Point<int, 2> ToIntFn(Point<float, 2> const& pt)
+Point<int, 2> ToIntFn(Point<double, 2> const& pt)
 {
    return Point<int, 2>(ToInt(pt.x),
                         ToInt(pt.y));
 }
 template <>
-Point<int, 3> ToIntFn(Point<float, 3> const& pt)
+Point<int, 3> ToIntFn(Point<double, 3> const& pt)
 {
    return Point<int, 3>(ToInt(pt.x),
                         ToInt(pt.y),
@@ -86,7 +86,7 @@ Point<int, 3> ToIntFn(Point<float, 3> const& pt)
 }
 
 template <int C>
-Point<int, C> csg::ToInt(Point<float, C> const& pt) {
+Point<int, C> csg::ToInt(Point<double, C> const& pt) {
    return ToIntFn(pt);
 }
 
@@ -96,22 +96,22 @@ Point<int, C> const& csg::ToInt(Point<int, C> const& pt) {
 }
 
 template <int C>
-Point<int, C> ToClosestIntFn(Point<float, C> const& pt);
+Point<int, C> ToClosestIntFn(Point<double, C> const& pt);
 
 template <>
-Point<int, 1> ToClosestIntFn(Point<float, 1> const& pt)
+Point<int, 1> ToClosestIntFn(Point<double, 1> const& pt)
 {
    return Point<int, 1>(ToClosestInt(pt.x));
 }
 
 template <>
-Point<int, 2> ToClosestIntFn(Point<float, 2> const& pt)
+Point<int, 2> ToClosestIntFn(Point<double, 2> const& pt)
 {
    return Point<int, 2>(ToClosestInt(pt.x),
                         ToClosestInt(pt.y));
 }
 template <>
-Point<int, 3> ToClosestIntFn(Point<float, 3> const& pt)
+Point<int, 3> ToClosestIntFn(Point<double, 3> const& pt)
 {
    return Point<int, 3>(ToClosestInt(pt.x),
                         ToClosestInt(pt.y),
@@ -119,7 +119,7 @@ Point<int, 3> ToClosestIntFn(Point<float, 3> const& pt)
 }
 
 template <int C>
-Point<int, C> csg::ToClosestInt(Point<float, C> const& pt) {
+Point<int, C> csg::ToClosestInt(Point<double, C> const& pt) {
    return ToClosestIntFn(pt);
 }
 
@@ -128,7 +128,7 @@ Point<int, C> const& csg::ToClosestInt(Point<int, C> const& pt) {
    return pt;
 }
 
-template<typename S> S csg::Interpolate(S const& p0, S const& p1, float alpha)
+template<typename S> S csg::Interpolate(S const& p0, S const& p1, double alpha)
 {
    if (alpha <= 0.0f) {
       return p0;
@@ -140,17 +140,17 @@ template<typename S> S csg::Interpolate(S const& p0, S const& p1, float alpha)
    return p0 + (delta * alpha);
 }
 
-template float csg::Interpolate(float const&, float const&, float);
+template double csg::Interpolate(double const&, double const&, double);
 
 #define DEFINE_POINT_CONVERSIONS(C) \
-   template Point<float, C> csg::ToFloat(Point<int, C> const&); \
-   template Point<float, C> const& csg::ToFloat(Point<float, C> const&); \
+   template Point<double, C> csg::ToFloat(Point<int, C> const&); \
+   template Point<double, C> const& csg::ToFloat(Point<double, C> const&); \
    template Point<int, C> const& csg::ToInt(Point<int, C> const&); \
-   template Point<int, C> csg::ToInt(Point<float, C> const&); \
+   template Point<int, C> csg::ToInt(Point<double, C> const&); \
    template Point<int, C> const& csg::ToClosestInt(Point<int, C> const&); \
-   template Point<int, C> csg::ToClosestInt(Point<float, C> const&); \
-   template Point<float, C> csg::Interpolate(Point<float, C> const&, Point<float, C> const&, float); \
-   template float csg::Interpolate(float const&, float const&, float);
+   template Point<int, C> csg::ToClosestInt(Point<double, C> const&); \
+   template Point<double, C> csg::Interpolate(Point<double, C> const&, Point<double, C> const&, double); \
+   template double csg::Interpolate(double const&, double const&, double);
 
 template <> Point3 csg::ConvertTo(Point3 const& pt) { return pt; }
 template <> Point3f csg::ConvertTo(Point3f const& pt) { return pt; }
