+++ source/csg/region.cpp
@@ -331,7 +331,7 @@ Point<S, C> Region<S, C>::GetClosestPoint(Point const& from) const
 {
    ASSERT(!IsEmpty());
 
-   float bestDistance = FLT_MAX;
+   double bestDistance = FLT_MAX;
 
    Point closest;
    for (int i = 0; i < C; i++) {
@@ -341,7 +341,7 @@ Point<S, C> Region<S, C>::GetClosestPoint(Point const& from) const
 
    for (const auto &cube : cubes_) {
       Point candidate = cube.GetClosestPoint(from);
-      float candidateDistance = from.SquaredDistanceTo(candidate);
+      double candidateDistance = from.SquaredDistanceTo(candidate);
       if (candidateDistance < bestDistance) {
          bestDistance = candidateDistance;
          closest = candidate;
@@ -571,9 +571,9 @@ void Region<S, C>::OptimizeOneTagByOctTree(S minCubeSize)
 
    for (int d = 0; d < C; d++) {
       // floor quantizes towards -infinity which is what we want
-      min[d] = static_cast<S>(std::floor((float)bounds.GetMin()[d] / cubeSize) * cubeSize);
+      min[d] = static_cast<S>(std::floor((double)bounds.GetMin()[d] / cubeSize) * cubeSize);
       // max[d] is an open interval and not included in the set
-      max[d] = static_cast<S>(std::ceil((float)bounds.GetMax()[d] / cubeSize) * cubeSize);
+      max[d] = static_cast<S>(std::ceil((double)bounds.GetMax()[d] / cubeSize) * cubeSize);
    }
 
    Cube quantizedBounds = Cube(min, max);
@@ -736,10 +736,10 @@ Region3 radiant::csg::GetBorderXZ(const Region3 &other)
 #endif
 
 template <int C>
-Region<float, C> csg::ToFloat(Region<int, C> const& region) {
+Region<double, C> csg::ToFloat(Region<int, C> const& region) {
    // xxx: how about a fast path that looks for cubes?  T(2n) usually
 
-   Region<float, C> result;
+   Region<double, C> result;
    for (Cube<int, C> const& cube : region) {
       result.Add(ToFloat(cube));    // make no be unique due to rounding!  see csg::ToInt(Cube)
    }
@@ -747,14 +747,14 @@ Region<float, C> csg::ToFloat(Region<int, C> const& region) {
 }
 
 template <int C>
-Region<float, C> const& csg::ToFloat(Region<float, C> const& region) {
+Region<double, C> const& csg::ToFloat(Region<double, C> const& region) {
    return region;
 }
 
 template <int C>
-Region<int, C> csg::ToInt(Region<float, C> const& region) {
+Region<int, C> csg::ToInt(Region<double, C> const& region) {
    Region<int, C> result;
-   for (Cube<float, C> const& cube : region) {
+   for (Cube<double, C> const& cube : region) {
       result.Add(ToInt(cube)); // so expensive!
    }
    return result;
@@ -788,19 +788,19 @@ void Region<S, C>::Validate() const
 }
 
 template <typename S, int C>
-Point<float, C> csg::GetCentroid(Region<S, C> const& region)
+Point<double, C> csg::GetCentroid(Region<S, C> const& region)
 {
-   Point<float, C> weightedSum = Point<float, C>::zero;
-   float totalWeight = 0;
+   Point<double, C> weightedSum = Point<double, C>::zero;
+   double totalWeight = 0;
 
    for (Cube<S, C> const& cube : region) {
-      Point<float, C> cubeCentroid = GetCentroid(cube);
-      float cubeArea = (float)cube.GetArea();
+      Point<double, C> cubeCentroid = GetCentroid(cube);
+      double cubeArea = (double)cube.GetArea();
       weightedSum += cubeCentroid.Scaled(cubeArea);
       totalWeight += cubeArea;
    }
 
-   Point<float, C> centroid = weightedSum.Scaled(1 / totalWeight);
+   Point<double, C> centroid = weightedSum.Scaled(1 / totalWeight);
    return centroid;
 }
 
@@ -848,10 +848,10 @@ MAKE_REGION(Region1)
 MAKE_REGION(Region1f)
 
 #define DEFINE_REGION_CONVERSIONS(C) \
-   template Region<float, C> csg::ToFloat(Region<int, C> const&); \
-   template Region<float, C> const& csg::ToFloat(Region<float, C> const&); \
+   template Region<double, C> csg::ToFloat(Region<int, C> const&); \
+   template Region<double, C> const& csg::ToFloat(Region<double, C> const&); \
    template Region<int, C> const& csg::ToInt(Region<int, C> const&); \
-   template Region<int, C> csg::ToInt(Region<float, C> const&); \
+   template Region<int, C> csg::ToInt(Region<double, C> const&); \
 
 DEFINE_REGION_CONVERSIONS(1)
 DEFINE_REGION_CONVERSIONS(2)
